Git scrub here. Penultimate merge for finished Patlak model.
