+++ Gpufit/constants.h
@@ -12,7 +12,11 @@ enum ModelID {
     LINEAR_1D = 5,
     FLETCHER_POWELL_HELIX = 6,
     BROWN_DENNIS = 7,
-	PATLAK = 12
+	PATLAK = 12,
+	TOFTS = 13,
+	TOFTS_EXTENDED = 14,
+	TISSUE_UPTAKE = 15,
+	TWO_COMPARTMENT_EXCHANGE = 16
 };
 
 // estimator ID
